Turned all sensors into class files
- created new class 'Sensors' with pure virtual functions common
  to all sensors implmented so far for this device
- turned all sensors (BME680, SFA30, MLX90614) into class files
  derived from new superclass 'Sensors'
- renamed global variable 'sensors' (sensorReadings_t) used to
  store all sensor readings to 'readings' to avoid confusion with
  new class 'Sensors'
- fixed getRuntimeMinutes() only adding up rounded number of seconds,
  which led to increasingly incorrect values for total runtime
- fixed dialog in wifi_init() previously ignoring the 'No' button
  and not skipping autoconnect() on 'Yes'
- splitted callback saveSetting() in wlan.cpp into saveParams() and
  saveWifi() fixing WDT not being (re)started when changing params
- added current time to optional memory monitor on serial console
- renamed bm680.cpp to bme680.cpp
- v1.5b3

+++ src/rtc.cpp
@@ -95,11 +95,11 @@ char* getDateString() {
 // returns runtime in minutes
 uint32_t getRuntimeMinutes() {
     static time_t lastMillis = 0;
-    static uint32_t seconds = 0;
+    static float seconds = 0.0;
 
-    seconds += tsDiff(lastMillis) / 1000;
+    seconds += tsDiff(lastMillis) / 1000.0;
     lastMillis = millis();
-    return seconds/60;
+    return (uint32_t)seconds/60;
 }
 
 
@@ -134,8 +134,3 @@ void ntp_init() {
     }
     delay(1500);
 }
-
-
-
-
-
